--- starter_code/app.js
@@ -1,26 +1,0 @@
-const express = require('express');
-const hbs = require('hbs');
-
-// require spotify-web-api-node package here:
-
-
-
-const app = express();
-
-app.set('view engine', 'hbs');
-app.set('views', __dirname + '/views');
-app.use(express.static(__dirname + '/public'));
-
-
-// setting the spotify-api goes here:
-
-
-
-
-
-
-// the routes go here:
-
-
-
-app.listen(3000, () => console.log("My Spotify project running on port 3000 🎧 🥁 🎸 🔊"));
