+++ views/index.hbs
@@ -1,0 +1,8 @@
+  <div class="home-container">
+  <div class="form-wrapper">
+    <form action="/artists" class="form-group home-form">
+      <input placeholder="Search for an artist!" class="form-control home-input" type="text" name="artist" id="artist-input">
+      <button class="btn btn-primary home-search" type="submit">Search for an Artist</button>
+    </form>
+  </div>
+</div>
